Incorporate newer samples
