+++ .gitignore
@@ -51,6 +51,7 @@ coverage.xml
 .hypothesis/
 .pytest_cache/
 cover/
+*.gradle/
 
 # Translations
 *.mo
